backend/requirements.txt
kafka-python
pandas
ta
scikit-learn==1.6.1
joblib
yfinance
